Enhancements and fixes to nos2xt
o Changed default behavior to skip updating a tide record if it already exists (can be overridden with '-u').

o Allow a "country code override" to be specified on command line.

o Allow a country code to be specified in import file

o Created an import file for all of Caribbean

o 'position' and 'datumOffset' fields were not being set correctly